Add LGPLv3 as a second possible license
With this move, which Carsten and I have planned for quite a while,
dune-foamgrid is now dual-licensed: you can either use the old
GPLv2 + linking exception license, or, alternatively, LGPLv3.
This licensing model is identical to the one used by PDELab.

Also, this patch adds the current list of copyright holders.