[growth] Major rewrite of the prototype grid growth interface
The grid growth interface is now completely independent of the adaptation interface.
The mark and getMark method are not overloaded anymore. Grid growth uses own mark functions
markForGrowth, markForRemoval, markForMerging that have easy readable names. The step by step
approach of preGrow, grow, postGrow remains the same. PreGrow returns true if elements are added.
grow returns true if elements or boundaries are added (I need to extrapolate the solution and/or
assign boundary conditions).

The internal state variables get more meaningful names. ADD_NEIGHBOR if a neighbor is added to
this element resulting in growth. VANISH if the entity vanishes. MERGE_WITH_NEIGHBOR if a new
element is created by connecting to existing facets.

This also implements the corresponding functions growSimplexElement, removeSimplexElement,
and mergeSimplexElement. The test is extended to show the new functionality and do some error checking.

Geometrical information needed for the growth is passed when marking and stored inside the element.