Make sure boundary intersections always have a valid boundarySegmentIndex.
Right now the boundarySegmentIndex is not a unique identifier to a certain facet.
The boundarySegmentIndex of a facet can change at growth/merge/removal.
This needs some thought. Unique boundaryIds might me more useful. This way the
user can write a map from boundaryIds (are sadly named boundarySegmentIndex in the intersection interface now),
to boundaryFlags/information on boundary conditions that always stays valid, and needs a new element
everytime a new boundary is created by growth or removal.