Example program solving the transport equation with a finite volume method
This is the initial commit of an example program solving the transport
equation using a finite volume method.  This first commit simply copies
an existing program using YaspGrid into the correct location.  Future
patches will replace the YaspGrid by a FoamGrid, and make it work
on an embedded network grid.