Fix IntersectionLeaf/LevelIterator
Introduce a null iterator factory to provide an end iterator for the intersection iterator.
Implement the leaf iterator completely analogously to the level iterator except it also finds
different level  neighbors. If the starting entity is not on the leaf this now correctly returns
an end iterator. This also simplifies the code (no extra obscure pushleaffacet function).