Add an example using parametrized elements
A triangle shaped one-dimensional network in a two dimensional world is
constructed where the elements are parametrized to map onto the unit circle.
Refinement shows that the grid gets closer to the unit cirlce with every refinement.

Then we solve a 1d heat equation using finite volume scheme on the finest circle grid