[fix] Publish new elements in the sons of the connecting facet if that is not leaf
The inserted element can be on a lower level then the the leaf versions of its facets.
As the new element is itself a leaf element it also has to be published in its facet's
sons if the facets are not on the leaf. This has to be done recursively until the leaf facet
also contains the element.