[test][examples] Make make install work again. Disable test magic.
We want to be able to install dune-foamgrid, not run all tests on
make all to reduce the runtime of dunecontrol, and disable the
test magic that never worked in foamgrid due to different naming
of the tests. This makes all of them work. Make all is available
in the subfolders examples/ and dune/foamgrid/test. Examples
and tests don't get installed.