[growth] Introduce a growthInsertionIndex for better user data control
Each element has a growthInsertionIndex that can be queried using a grid method
with the same name. This index is only valid between calling grow() and postGrow()
and only for elements that have been inserted using the grids insertElement method.
It can be used to attach user data to an element inserted with the insertElement method,
because we don't know at insertion time (before calling grow()) which index or id
the element will have in the grid. The concept is very similar to the insertionIndex
of the factory with the difference that the validity of the index is only temporary.