[growth] Fix double vertex insertion
When inserting two element that share a new vertex this vertex was inserted twice
resulting in a disconnected grid. The check if a vertex is already inserted was
already in place, however the map had the wrong scope and the pointer written
in the map was a pointer to the grid entity not the entity in the to-be-inserted queue.