[foamgrid] Initialize entity vector and level 0
A default constructed FoamGrid should be a valid (but empty) grid.
The list of entites was previously uninitialized but this data member
is accessed in several interface functions.

The entity list is now properly initialized. Level 0 always exists, but is
empty (empty entity lists) when the grid is default constructed.
An empty level index set is also created for level 0.